fix after test

+++ FaceRec/order.py
@@ -10,6 +10,8 @@ import face_recognition
 import os
 import time
 import database
+import cv2
+import numpy as np
 
 def facecompare(klas, periode):
     datetime = time.strftime("%c")
@@ -27,22 +29,36 @@ def facecompare(klas, periode):
             print("new student made")
     else:
         for filename in os.listdir(temp):
-            frank_image = face_recognition.load_image_file(temp + filename)
-            frank_face_encoding = face_recognition.face_encodings(frank_image)[0]
+            unknown_image = face_recognition.load_image_file(temp + filename)
+            unknown_face_encoding = face_recognition.face_encodings(unknown_image)[0]
             searching = True
             i = 0
             while i < len(dirs) and searching:
                 dirs2 = os.listdir(folder + dirs[i])
-                unknown_image = face_recognition.load_image_file(folder + dirs[i] + "/" + dirs2[0])
-                unknown_face_encoding = face_recognition.face_encodings(unknown_image)[0]
+                index = 0
+                size = 0
+                count = 0
+                for file in dirs2:
+                    im = cv2.imread(folder + dirs[i] + "/" + file)
+                    height = np.size(im, 0)
+                    width = np.size(im, 1)
+                    tsize = width + height
+                    if tsize > size:
+                        size = tsize
+                        count = index
+                    index+=1
+                print(dirs2[count])
+                known_image = face_recognition.load_image_file(folder + dirs[i] + "/" + dirs2[count])
+                known_face_encoding = face_recognition.face_encodings(known_image)[0]
                 known_faces = [
-                     frank_face_encoding
+                     known_face_encoding
                 ] 
                 results = face_recognition.compare_faces(known_faces, unknown_face_encoding)
                 
                 if results[0]:
-                    os.rename(temp + filename, folder + dirs[i] + "/" + datetime)
-                    database.AddPresence(dirs[i], klas, periode)
+                    if not os.path.exists(folder + dirs[i] + "/" + datetime):
+                        os.rename(temp + filename, folder + dirs[i] + "/" + datetime)
+                        database.AddPresence(dirs[i], klas, periode)
                     searching = False
                 else:
                     i+=1
